requirements.txt
-e .
aws-cdk.aws-s3
aws-cdk.aws-iam